Juan

+++ database/migrations/2024_10_12_160802_create_municipios_table.php
@@ -13,14 +13,15 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::dropIfExists('municipio');
-        Schema::create('municipio', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('id_departamento')->constrained('departamento')->onUpdate('cascade')->onDelete('cascade');
-            $table->string('nombre_municipios');
-            $table->integer('codigo_municipios');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('municipio')) {
+            Schema::create('municipio', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('id_departamento')->constrained('departamento')->onUpdate('cascade')->onDelete('cascade');
+                $table->string('nombre_municipios');
+                $table->integer('codigo_municipios');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
